Initial commit
Fast CGI code originally from fastcgi++ (http://fastcgipp.com/)

Much of this was originally written in python, but it turned out
that my firewall is too slow to run python fastcgi apps.  So I
set out to use c++ instead.

The theme was based on the drupal 5 garland theme, modified to
be the flames theme.  It is a nearly all CSS based, so easily
changeable.