modify json queries so any number of forms or records can be requested
calling postJSON with data like this will encode things properly:
{ static_dhcp: "records", dhcp_settings: "form" }

Also, this set up the variable model so it can retrieve settings
for each form to be sent back with the form elements.

Oh... and by the way this commit is going to have to record the
veracity of origin for the rest of the commits since I forgot.
All the code herein was either written by me, or came from a
project that had a GPL compatible license and is now subject to
the license found in the LICENSE file.

Signed-off-by: Vernon Mauery <vernon@mauery.com>