Fixed memory corruption in http.cpp
http.cpp had a couple of extra deletes to a pointer that was already
assigned in a scoped array.  Also, at the end, it reallocates a new
buffer for the scoped array that is a subset of the current buffer, but
it freed the old memory before referencing it.  Bad juju.

To validate this, I also had some minor improvements to my malloc/free
debug checker which are included in this commit.

Signed-off-by: Vernon Mauery <vernon@mauery.com>