Add a config file and parser
Until now, things like where the json form dir is and where the
database lives were all hard-coded values.  Now, one can set the
values in /etc/pyrobox.conf.

After adding the conf file code, I modified the logger, the model
class and the json form code to use it.

I also found a bug in the rtrim/trim code that I fixed in this
commit.  It was truncating the last character of the string.

Signed-off-by: Vernon Mauery <vernon@mauery.com>