Move form render, submit, commit to a form class
This change pulls the form render, sumbit, and commit phases
out of config.cpp into a specialized form class.  In time this
class will be able to validate the submitted values as well.

Note that the current plan is to get rid of the forms stored as
JSON text and replace them with some other configuration medium.
That step is still under debate.  The reason for moving away from
the JSON representation of the form is that it makes it harder
for the form to be validated on the server.  However, I really
dislike the notion of having each form hard coded into the
application.   This makes for a very slow development cycle.
I would prefer some non-C++ representation of the form that
can be easily parsed to create a form object that can be rendered
as JSON and sent to the browser, and easily validated.

Suggestions are welcome. :)

Signed-off-by: Vernon Mauery <vernon@mauery.com>