Fix checkbox form element in a couple of ways
Checkbox form elements were fundamentally broken before this commit.
I forget how messed up checkboxes are in html land.  When checked, they
post a value.  When unchecked, they don't post anything.

I had to fix http.cpp's handling of empty values such as
a=1&b=2&c=&d=4
c= was causing a segfault.

Then I had to allow a new attribute of a checkbox.  This is the
value_checked attribute, or the value that will be posted if the
checkbox is checked.  If the value or default value of the variable
is the same as value_checked, the checkbox renders as being checked.

Finally, I had to check the values getting posted.  If the form item
is a checkbox and no value of that name has been posted, add an empty
value to the list of values posted.

Signed-off-by: Vernon Mauery <vernon@mauery.com>