Add port forwarding page
Add the port forwarding page that has the table showing current
port forwards and a link to add a new one.  Using the generic
tables infrastructure this was not a huge deal.  The work went
along these lines:

1. Add the model class to model.hpp/model.cpp
2. Add the model class to the two factories in model.cpp
3. Create a form that references the model
4. Add pyrobox.js and a call to json_forms_and_tables
   that references the form
5. Add the table to the html page that is referenced by the form
6. Wrap the table with a popup add class if desired
7. Fix any bugs that appear and do regression testing on
   existing tables to make sure changes didn't break those too.

This is the path that I followed to add this page.  In the
bug fixing department, here is a laundry list of what I found:

1. Not all items are inline-form ready.  I had to modify the
   select form item in pyrobox.js to render properly.
2. db.cpp was not properly handling null fields in the database
3. Table sorter does not appear to like empty tables. (not fixed)
4. The add new row logic does not accomodate empty tables. (not fixed)

Signed-off-by: Vernon Mauery <vernon@mauery.com>