Move form to use json_forms_and_tables function
Pages created before settings/dhcp.html used a hand-crafted method
for fetching the form and setting up the ajax submit routine.  This
patch updates those pages to use the common json_forms_and_tables
function.  The only page that still uses hand-crafted code is the
network/wan page because it has a special type of form on it.

Signed-off-by: Vernon Mauery <vernon@mauery.com>